feat: Add new game button to win screen
Gemini commit: Can you make a way to test the "you win!" message without winning a game? let me press "w" to show it and then "w" to hide it again (except at the real end of a game...) after that, let's add a "play a new game" button to the "you win!" message. do those as two separate commits