refactor(animdemo): Use Canvas API for animation
This change refactors the animdemo service to use the Canvas 2D API
instead of SVG for rendering the animation. This provides a significant
performance improvement, as the Canvas API is better suited for drawing
a large number of simple shapes.

Gemini commit: is there some other way besides svg to do this? how about drawing on a canvas instead of adding path to an svg? it's just lines.