refactor(js): abstract syscall/js details from view logic
This commit refactors the thegrid/js package to provide a more complete abstraction over the low-level `syscall/js` package. This removes the need for application-level code, such as the explorers view, to have any knowledge of the underlying JS interop details.

The following changes were made:

-   `IsNull()` and `Equal()` methods were added to the `DOMElement` type.
-   A `Null()` function was added to the `js` package to create a null `DOMElement`.
-   The explorers view was refactored to use this new, cleaner API, removing all direct dependencies on `syscall/js`.

Gemini commit: I don't like that view.go is referencing syscall/js at all - it's an indication that we're missing some vital abstractions on the thegrid/js package API. what are we using it for, and what would it look like to fix that?