Refactor Firebase auth to use modular SDK
Replaced the legacy Firebase compat scripts with the modern modular SDK.
This change adapts the existing Go-based JavaScript abstraction to work with the new SDK.

- Updated `thegrid/firebase/auth/auth.js` to use ES modules.
- Installed `firebase` and `esbuild` npm packages.
- Updated `thegrid/package.json` to build the modular JS code.
- Updated `thegrid/apps/delta/index.html` to use the new bundled JS.

Gemini commit: we were looking at firebase in the js framework in the thegrid directory, and i was going to ask further questions on what you wrote in the last git commit. i notice that the firebase auth is dependent on a compat.js file in thegrid/apps/delta/index.html, but i asked you about using real "modules"... you started to run some npm commands, is that really the right way to go?