fix: Add templates to server data dependencies
The server was failing to parse the layout templates because they were not included in the server's data dependencies. This change adds the template files to the `data` attribute of the `go_binary` rule in `server/BUILD.bazel`.

Gemini commit: that's better. the error is:2025/08/16 22:31:54 failed to parse layout templates: open server/templates/layout.html.tpl: no such file or directory